add serde serialization for String
Additionally reorganize tests to split the ones
that depend on `collections` and `boxed` features.
